requirements.txt
# Web Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
aiofiles>=23.0.0

# Config & Validation
pydantic>=2.7.0
pydantic-settings>=2.2.0

# AI & NLP (Privacy Layer)
spacy>=3.7.0

# Agentic Framework & LLM
groq>=0.9.0

# Vector Database (Memory & Hybrid Search)
pinecone>=3.2.0
sentence-transformers>=2.7.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0

# Production Logging
python-json-logger>=2.0.7
